style(msvc): fix C4244/C4267 in every remaining 2-site file (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226; [NotesLoader],[NoteData]
-tagged tests (646/46) re-checked identical.

26 files, 52 sites: StepsUtil.cpp, StageStats.cpp, Sprite.cpp,
ScrollBar.cpp, ScreenSelectMusic.cpp, ScreenOptionsManageProfiles.cpp,
ScreenMapControllers.cpp, ScreenInstallOverlay.cpp, SampleHistory.cpp,
RandomSample.cpp, RageUtil_CharConversions.cpp,
RageSoundReader_SpeedChange.cpp, RageSoundReader_Resample_Good.cpp,
RageSoundReader_Chain.cpp, RageModelGeometry.cpp,
RageFileDriverTimeout.cpp, ProfileManager.cpp, PlayerOptions.cpp,
Player.cpp, NotesLoaderSMA.cpp (SS5-protected, characterization-only),
NotesLoaderJson.cpp (SS5-protected, characterization-only),
NoteData.cpp (SS5-protected NoteData*, characterization-only),
MsdFile.h, GameSoundManager.cpp, CsvFile.cpp, CharacterManager.cpp.

Same recurring narrowing patterns throughout: size_t/double-family
values narrowed at wrap()/RandomInt() calls, iterator-diffs,
reverse-loop bounds, RString::Left/Right length args, and Lua
lua_push*/lua_raw* bindings.

+++ src/GameSoundManager.cpp
@@ -917,11 +917,11 @@ int LuaFunc_get_sound_driver_list(lua_State* L)
 {
 	std::vector<RString> driver_names;
 	split(RageSoundDriver::GetDefaultSoundDriverList(), ",", driver_names, true);
-	lua_createtable(L, driver_names.size(), 0);
+	lua_createtable(L, static_cast<int>(driver_names.size()), 0);
 	for(std::size_t n= 0; n < driver_names.size(); ++n)
 	{
 		lua_pushstring(L, driver_names[n].c_str());
-		lua_rawseti(L, -2, n+1);
+		lua_rawseti(L, -2, static_cast<int>(n+1));
 	}
 	return 1;
 }
