+++ src/RageUtil_CharConversions.cpp
@@ -14,7 +14,7 @@
 /* Convert from the given codepage to UTF-8.  Return true if successful. */
 static bool CodePageConvert( RString &sText, int iCodePage )
 {
-	int iSize = MultiByteToWideChar( iCodePage, MB_ERR_INVALID_CHARS, sText.data(), sText.size(), nullptr, 0 );
+	int iSize = MultiByteToWideChar( iCodePage, MB_ERR_INVALID_CHARS, sText.data(), static_cast<int>(sText.size()), nullptr, 0 );
 	if( iSize == 0 )
 	{
 		LOG->Trace( "%s\n", werr_ssprintf(GetLastError(), "err: ").c_str() );
@@ -24,7 +24,7 @@ static bool CodePageConvert( RString &sText, int iCodePage )
 	std::wstring sOut;
 	sOut.append( iSize, ' ' );
 	/* Nonportable: */
-	iSize = MultiByteToWideChar( iCodePage, MB_ERR_INVALID_CHARS, sText.data(), sText.size(), (wchar_t *) sOut.data(), iSize );
+	iSize = MultiByteToWideChar( iCodePage, MB_ERR_INVALID_CHARS, sText.data(), static_cast<int>(sText.size()), (wchar_t *) sOut.data(), iSize );
 	ASSERT( iSize != 0 );
 
 	sText = WStringToRString( sOut );
